requirements.txt
# requirements.txt

# Core ML
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.40.0
peft>=0.10.0
bitsandbytes>=0.43.0
accelerate>=0.29.0
sentencepiece>=0.1.99
protobuf>=3.20.0

# Vision
Pillow>=10.0.0
open-clip-torch>=2.24.0

# Training utilities
datasets>=2.18.0
tqdm>=4.66.0

# Server
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic>=2.0.0

# Frontend
streamlit>=1.33.0
requests>=2.31.0
